Refactoring of Advanced Image Adapter
