+++ src/main/java/cec/input/Data.java
@@ -66,7 +66,8 @@ public class Data {
         TEXT_SPACE("text/space-separated-values", "Text file with real values separated by a single space character."),
         TEXT_CSV("text/comma-separated-values", "Text file with real values separated by a single comma character. Note that some locales use the comma character as a decimal point while the CSV format separator is a semicolon character. Thic CSV format is currently not supported."),
         IMAGE_PNG("image/png", "A PNG file. Maps pixels to points in a two-dimensional space with point weights proportional to the pixels' shades in grayscale."),
-    	IMAGE_TIFF("image/tiff", "A TIFF file.");
+    	IMAGE_TIFF("image/tiff", "A TIFF file."),
+    	IMAGE_BMP("image/bmp", "A BMP file");
     
 
         private final String description;
